Improve detection of new turns
I don't understand all the edge cases, but in my testing this new
algorithm seems to work very well. It watches the queue of next pairs
(including the currently falling one) for changes. After it's changed,
it looks for a state where the current falling pair is at position (2,
11). As an additional check, the currently falling pair must be what we
expect from looking at the queue in the past.